test: split profile wifi entry coverage
# Conflicts:
#	pubspec.yaml
