+++ pubspec.yaml
@@ -55,6 +55,7 @@ dev_dependencies:
   flutter_lints: ^6.0.0
   build_runner: ^2.13.1
   pointycastle: ^4.0.0
+  shared_preferences_platform_interface: ^2.4.1
   slang_mcp: ^0.1.3
 
 flutter:
